fix: archer arrows spawn/despawn with character
This fix addresses the issue of arrows not despawning when the archer respawns. The archer class generates arrows upon spawn, but would not despawn them when respawning/exiting the game, resulting in unused arrows sitting in memory, multiplied by the number of respawns of an archer character. By spawning arrows within a container gameobject, and attaching that container to the archer gameobject, the arrows can despawn along with the archer gameobject.

Arrow.prefab
- Set arrow to deactivated by default

ObjectPool.cs
- Created container gameobject field and updated the constructor to make arrows within the container, as well as make the container a child of the gameobject creating the object pool

ArcherActions.cs
- Updated pool to reflect updates to its constructor
- Removed ondisable